Move BoundingBox to its own header
We'd like to use it outside of MeshTools too.
