+++ src/libmesh_SOURCES
@@ -179,6 +179,7 @@ libmesh_SOURCES =  \
         src/mesh/abaqus_io.C \
         src/mesh/boundary_info.C \
         src/mesh/boundary_mesh.C \
+        src/mesh/bounding_box.C \
         src/mesh/checkpoint_io.C \
         src/mesh/distributed_mesh.C \
         src/mesh/ensight_io.C \
